moved handler struct and relevant code to handler.rs
